Issues 1179 and 1182 (#1232)
* Update logging

* Remove duplicate code

* Update unit tests

* Clean up includes

* Merge ProtobufSerializer into protobuf_util namespace

* Precompile regex

* Add const

* Add Split() convenience method, update log level/ID parsing

* Move log.h to legacy folder

* Elimininate gotos

* Fixes for gcc 13

* Update compiler flags

* Update default folder handling

* Use references instead of pointers

* Move code for better encapsulation

* Move code

* Remove unused method argument

* Move device logger

* Remove redundant to_string

* Rename for consistency

* Update handling of protobuf pointers

* Simplify protobuf usage

* Memory handling update

* Add hasher

+++ cpp/scsiloop/scsiloop_timer.cpp
@@ -10,7 +10,7 @@
 #include "scsiloop_timer.h"
 #include "hal/systimer.h"
 #include "scsiloop/scsiloop_cout.h"
-#include "shared/log.h"
+#include "hal/log.h"
 
 int ScsiLoop_Timer::RunTimerTest(vector<string> &error_list)
 {
@@ -56,7 +56,7 @@ int ScsiLoop_Timer::RunTimerTest(vector<string> &error_list)
     }
     after            = SysTimer::GetTimerLow();
     elapsed_nanosecs = after - before;
-    LOGDEBUG("SysTimer::SleepUsec() Average %d", (uint32_t)(elapsed_nanosecs / 100));
+    LOGDEBUG("SysTimer::SleepUsec() Average %d", elapsed_nanosecs / 100);
 
     if ((elapsed_nanosecs > expected_usec_result * (1.0 + timer_tolerance_percent)) ||
         (elapsed_nanosecs < expected_usec_result * (1.0 - timer_tolerance_percent))) {
